fix(anland): frame_callback_sequence is u32 not u64
Compiler error in GĐ2 commit (8ba80b24): pending_cb_sequence is u32
from output_state.frame_callback_sequence, not u64.

+++ src/backend/anland.rs
@@ -702,7 +702,7 @@ impl Anland {
         &mut self,
         niri: &mut Niri,
         output: &Output,
-        _cb_sequence: u64,
+        _cb_sequence: u32,
     ) {
         if let Some(token) = self.deferred_callback_token.take() {
             let _ = niri.event_loop.remove(token);
